Refactor Structure to have a base module for ColumnTypes and JuBaseClasses
