+++ julearn/transformers/__init__.py
@@ -15,4 +15,3 @@ from . dataframe import (
 )
 from . target import TargetTransfromerWrapper
 from . cbpm import CBPM
-from . base import JuTransformer
