+++ julearn/utils/__init__.py
@@ -5,9 +5,3 @@ from . import logging
 
 from . logging import logger, configure_logging, raise_error, warn
 
-from . column_types import (
-    # pick_columns,
-    change_column_type,
-    get_column_type,
-    make_type_selector
-)
